[mod/#44] HapHapCategoryChipList 파일 위치 수정
