+++ app/src/main/java/com/haphap/app/presentation/common/component/HapHapCategoryChipList.kt
@@ -1,4 +1,4 @@
-package com.haphap.app.core.designsystem.component.chip
+package com.haphap.app.presentation.common.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -15,11 +14,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.haphap.app.R
+import com.haphap.app.core.designsystem.component.chip.FilterChipContent
+import com.haphap.app.core.designsystem.component.chip.HapHapFilterChip
 import com.haphap.app.core.designsystem.theme.HapHapTheme
 import com.haphap.app.data.model.search.ChipListModel
+import com.haphap.app.presentation.common.state.CategoryChipState
 import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.PersistentList
-import kotlinx.collections.immutable.persistentListOf
 
 @Composable
 fun HapHapCategoryChipList(
@@ -27,6 +27,7 @@ fun HapHapCategoryChipList(
     selectedChips: ImmutableList<String>,
     onFilterClick: (String) -> Unit,
     modifier: Modifier = Modifier,
+    onIconFilterClick: () -> Unit = {},
 ) {
     LazyRow(
         modifier = modifier.fillMaxWidth(),
@@ -38,7 +39,7 @@ fun HapHapCategoryChipList(
                 content = FilterChipContent.IconContent(
                     iconRes = R.drawable.ic_filter_20,
                 ),
-                onFilterClick = {},
+                onFilterClick = onIconFilterClick,
                 isFilterSelected = false
             )
         }
@@ -53,35 +54,6 @@ fun HapHapCategoryChipList(
                 isFilterSelected = selectedChips.contains(it.category),
             )
         }
-    }
-}
-
-
-@Immutable
-data class CategoryChipState(
-    val chipList: ImmutableList<ChipListModel> = DEFAULT_CHIP_LIST,
-    val selectedChips: PersistentList<String> = persistentListOf("전체"),
-) {
-    fun toggle(category: String): CategoryChipState =
-        copy(
-            selectedChips = when {
-                category == "전체" -> persistentListOf("전체")
-                selectedChips.size == 1 && selectedChips.contains(category) -> selectedChips
-                selectedChips.contains(category) -> selectedChips.remove(category)
-                else -> selectedChips.remove("전체").add(category)
-            }
-        )
-
-    companion object {
-        val DEFAULT_CHIP_LIST: ImmutableList<ChipListModel> = persistentListOf(
-            ChipListModel(id = 1, category = "전체"),
-            ChipListModel(id = 2, category = "기획"),
-            ChipListModel(id = 3, category = "마케팅/홍보"),
-            ChipListModel(id = 4, category = "인사"),
-            ChipListModel(id = 5, category = "영업"),
-            ChipListModel(id = 6, category = "개발/데이터"),
-            ChipListModel(id = 7, category = "금융/보험"),
-        )
     }
 }
