+++ app/src/main/java/com/haphap/app/presentation/joblist/JobListContract.kt
@@ -1,8 +1,8 @@
 package com.haphap.app.presentation.joblist
 
 import androidx.compose.runtime.Immutable
-import com.haphap.app.core.designsystem.component.chip.CategoryChipState
 import com.haphap.app.data.model.list.JobItemModel
+import com.haphap.app.presentation.common.state.CategoryChipState
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
